Updated tests to work with 2018.2
