+++ TestBehaviours/HierarchyBehaviour.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine;
+
+namespace DUCK.HieriarchyBehaviour.TestBehaviours
+{
+	[AddComponentMenu("")]
+	public class HierarchyBehaviour : MonoBehaviour, IHierarchyBehaviour
+	{
+		public bool DidInitialize { get; private set; }
+
+		public void Initialize()
+		{
+			DidInitialize = true;
+		}
+	}
+}
